requirements.txt
zeep
requests
configargparse
urllib3
pprint
requests_cache
arrow
bravado
elasticsearch
mysql-connector
